Move settings hook to utility file
